+++ packages/crypto/src/builder/index.ts
@@ -83,4 +83,15 @@ export class TransactionBuilderDirector {
 }
 
 const transactionBuilder = new TransactionBuilderDirector();
-export { transactionBuilder };
+export {
+    transactionBuilder,
+    DelegateRegistrationBuilder,
+    DelegateResignationBuilder,
+    IPFSBuilder,
+    MultiPaymentBuilder,
+    MultiSignatureBuilder,
+    SecondSignatureBuilder,
+    TimelockTransferBuilder,
+    TransferBuilder,
+    VoteBuilder,
+};
