+++ packages/crypto/src/builder/transactions/timelock-transfer.ts
@@ -1,5 +1,5 @@
 import { TransactionTypes } from "../../constants";
-import { feeManager } from "../../managers/fee";
+import { feeManager } from "../../managers";
 import { TransactionBuilder } from "./transaction";
 
 export class TimelockTransferBuilder extends TransactionBuilder {
