+++ packages/crypto/src/models/transaction.ts
@@ -5,7 +5,7 @@ import ByteBuffer from "bytebuffer";
 import { createHash } from "crypto";
 import { TransactionTypes } from "../constants";
 import { crypto } from "../crypto/crypto";
-import { configManager } from "../managers/config";
+import { configManager } from "../managers";
 import { Bignum } from "../utils";
 
 const { transactionIdFixTable } = configManager.getPreset("mainnet").exceptions;
